split main.go by some functions
